--- tests/ballerina-test/src/test/resources/test-src/statements/services/nativeimpl/outbound/request/out-request-native-function-negative.bal
@@ -1,75 +1,0 @@
-import ballerina.net.http;
-import ballerina.mime;
-
-function testGetContentLength (http:Request req) (int) {
-    int length = req.getContentLength();
-    return length;
-}
-
-function testGetHeader (http:Request req, string key) (string) {
-    var contentType = req.getHeader(key);
-    if (contentType == null) {
-        return null;
-    }
-    return contentType;
-}
-
-function testGetJsonPayload (http:Request req) (json, mime:EntityError) {
-    return req.getJsonPayload();
-}
-
-function testGetProperty (http:Request req, string propertyName) (string) {
-    string payload = req.getProperty(propertyName);
-    return payload;
-}
-
-function testGetEntity (http:Request req) (mime:Entity, mime:EntityError) {
-    return req.getEntity();
-}
-
-function testGetStringPayload (http:Request req) (string, mime:EntityError) {
-    return req.getStringPayload();
-}
-
-function testGetBinaryPayload (http:Request req) (blob, mime:EntityError) {
-    return req.getBinaryPayload();
-}
-
-function testGetXmlPayload (http:Request req) (xml, mime:EntityError) {
-    return req.getXmlPayload();
-}
-
-function testRemoveHeader (http:Request req, string key) (http:Request) {
-    req.removeHeader(key);
-    return req;
-}
-
-function testRemoveAllHeaders (http:Request req) (http:Request) {
-    req.removeAllHeaders();
-    return req;
-}
-
-function testSetHeader (http:Request req, string key, string value) (http:Request) {
-    req.setHeader(key, value);
-    return req;
-}
-
-function testSetJsonPayload (http:Request req, json value) (http:Request) {
-    req.setJsonPayload(value);
-    return req;
-}
-
-function testSetProperty (http:Request req, string name, string value) (http:Request) {
-    req.setProperty(name, value);
-    return req;
-}
-
-function testSetStringPayload (http:Request req, string value) (http:Request) {
-    req.setStringPayload(value);
-    return req;
-}
-
-function testSetXmlPayload (http:Request req, xml value) (http:Request) {
-    req.setXmlPayload(value);
-    return req;
-}
