--- tests/ballerina-test/src/test/resources/test-src/statements/services/nativeimpl/outbound/response/out-response-compile-negative.bal
@@ -1,12 +1,0 @@
-import ballerina.net.http;
-
-function testResponseSetStatusCode (http:Response res, string statusCode) (http:Response) {
-    res.statusCode = statusCode;
-    return res;
-}
-
-function testResponseGetMethod (http:Response res) (string) {
-    string method = res.method;
-    return method;
-}
-
